Write file, zip, and send to Datadog
