+++ src/commands/lambda/__tests__/functions/commons.test.ts
@@ -12,12 +12,11 @@ import {fromNodeProviderChain} from '@aws-sdk/credential-providers'
 import {mockClient} from 'aws-sdk-client-mock'
 
 import 'aws-sdk-client-mock-jest'
-import {CI_API_KEY_ENV_VAR} from '../../../../constants'
+import {CI_API_KEY_ENV_VAR, CI_SITE_ENV_VAR} from '../../../../constants'
 
 import {
   CI_API_KEY_SECRET_ARN_ENV_VAR,
   CI_KMS_API_KEY_ENV_VAR,
-  CI_SITE_ENV_VAR,
   DD_LAMBDA_EXTENSION_LAYER_NAME,
   DEFAULT_LAYER_AWS_ACCOUNT,
   EXTENSION_LAYER_KEY,
