+++ src/commands/lambda/__tests__/functions/uninstrument.test.ts
@@ -6,7 +6,7 @@ import {LambdaClient, ListFunctionsCommand} from '@aws-sdk/client-lambda'
 import {mockClient} from 'aws-sdk-client-mock'
 
 import 'aws-sdk-client-mock-jest'
-import {API_KEY_ENV_VAR} from '../../../../constants'
+import {API_KEY_ENV_VAR, SITE_ENV_VAR} from '../../../../constants'
 
 import {
   API_KEY_SECRET_ARN_ENV_VAR,
@@ -22,7 +22,6 @@ import {
   PROFILER_ENV_VAR,
   PROFILER_PATH_ENV_VAR,
   SERVICE_ENV_VAR,
-  SITE_ENV_VAR,
   SUBSCRIPTION_FILTER_NAME,
   TRACE_ENABLED_ENV_VAR,
   VERSION_ENV_VAR,
