+++ src/commands/lambda/__tests__/prompt.test.ts
@@ -1,7 +1,7 @@
 jest.mock('inquirer')
 import {prompt} from 'inquirer'
 
-import {CI_API_KEY_ENV_VAR} from '../../../constants'
+import {CI_API_KEY_ENV_VAR, CI_SITE_ENV_VAR} from '../../../constants'
 
 import {
   AWS_ACCESS_KEY_ID_ENV_VAR,
@@ -9,7 +9,6 @@ import {
   AWS_SESSION_TOKEN_ENV_VAR,
   CI_API_KEY_SECRET_ARN_ENV_VAR,
   CI_KMS_API_KEY_ENV_VAR,
-  CI_SITE_ENV_VAR,
 } from '../constants'
 import {
   confirmationQuestion,
