+++ src/helpers/fileSystem.ts
@@ -1,4 +1,7 @@
 import fs from 'fs'
+import path from 'path'
+
+import JSZip from 'jszip'
 
 /**
  * Delete a folder and all its contents
@@ -11,6 +14,66 @@ export const deleteFolder = (folderPath: string) => {
   } catch (err) {
     if (err instanceof Error) {
       throw Error(`Failed to delete files located at ${folderPath}: ${err.message}`)
+    }
+  }
+}
+
+/**
+ * Write the data to a file
+ * @param filePath path to the file
+ * @param data the data to write
+ * @throws Error if the file cannot be written
+ */
+export const writeFile = (filePath: string, data: string) => {
+  try {
+    fs.writeFileSync(filePath, data)
+  } catch (err) {
+    if (err instanceof Error) {
+      throw Error(`Unable to write file: ${err.message}`)
+    }
+  }
+}
+
+/**
+ * Zip the entire contents of a folder
+ * @param rootFolderPath path to the root folder to zip
+ * @param zipPath path to save the zip file
+ * @throws Error if the zip fails
+ */
+export const zipContents = async (rootFolderPath: string, zipPath: string) => {
+  const zip = new JSZip()
+
+  const addFolderToZip = (folderPath: string) => {
+    if (!fs.existsSync(folderPath)) {
+      throw Error(`Folder does not exist: ${folderPath}`)
+    }
+
+    const folder = fs.statSync(folderPath)
+    if (!folder.isDirectory()) {
+      throw Error(`Path is not a directory: ${folderPath}`)
+    }
+
+    const contents = fs.readdirSync(folderPath)
+    for (const item of contents) {
+      const fullPath = path.join(folderPath, item)
+      const file = fs.statSync(fullPath)
+
+      if (file.isDirectory()) {
+        addFolderToZip(fullPath)
+      } else {
+        const data = fs.readFileSync(fullPath)
+        zip.file(path.relative(rootFolderPath, fullPath), data)
+      }
+    }
+  }
+
+  try {
+    addFolderToZip(rootFolderPath)
+    const zipContent = await zip.generateAsync({type: 'nodebuffer'})
+    fs.writeFileSync(zipPath, zipContent)
+  } catch (err) {
+    if (err instanceof Error) {
+      throw Error(`Unable to zip files: ${err.message}`)
     }
   }
 }
